Move Tabs under FormTabs
